add run command and make it default
